adds assertion so it prevents moving same category to same parent id and to same child categories
