+++ src/PrestaShopBundle/Resources/views/Admin/Improve/Design/Cms/Blocks/category_form.html.twig
@@ -66,7 +66,7 @@
         }) }}
 
         {{ ps.form_group_row(cmsPageCategoryForm.meta_keywords, {}, {
-          'label': 'Meta description'|trans({}, 'Admin.Global'),
+          'label': 'Meta keywords'|trans({}, 'Admin.Global'),
           'help': invalidCharactersLabel
         }) }}
